requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
anndata==0.10.7
captum==0.7.0
numpy==1.24.3
datatable==1.1.0
#dgl==2.1.0+cu121
episcanpy==0.3.1
pandas==2.2.2
scanpy==1.10.1
scikit-learn==1.4.2
scipy==1.13.0
#torch==2.2.1+cu121
torch-geometric==2.5.2
#torchaudio==2.2.1+cu121
torchdata==0.7.1
#torchvision==0.17.1+cu121
tqdm==4.66.2
pyyaml==6.0.1
audtorch==0.6.4
pydantic==2.7.0
torch-sparse==0.6.18
torch-scatter==2.1.2
tensorboardx==2.6.2.2
